Revert "ADD respawn feature in tutorial map (#49)"
This reverts commit 4e4e39bff14983ec059809311e6ccbcd089c071c.
