--- SDP_Main/Assets/InitTestScene638201054136290159.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 952d8987b6f74224c88a8f58ff577fe7
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
