--- SDP_Main/Assets/Scripts/Entities/Player/Tests.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 2603fea57cf4a874b816d7f712ffa48b
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
