--- SDP_Main/Assets/Scripts/Entities/Player/Tests/TestScript.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8d66d3c01bf120b498de5e7826eafbbb
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
