--- SDP_Main/Assets/Scripts/UI/InGame/Toturial/Trigger1.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: bfd0c39e66411ac429eba68804acd89f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
