--- SDP_Main/Assets/Scripts/UI/InGame/Toturial/Trigger2.cs
@@ -1,34 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-using UnityEngine.UI;
-using TMPro;
-
-public class Trigger2 : MonoBehaviour
-{
-    public TextMeshProUGUI MoveText;
-    public TextMeshProUGUI SprintText;
-    public TextMeshProUGUI JumpText;
-
-    // public GameObject trigger2;
-
-    private void OnTriggerEnter(Collider other)
-    {
-        Debug.Log("Trigger work");
-        if (other.CompareTag("Player"))
-        {
-            //Move downward the tip text
-            MoveText.rectTransform.anchoredPosition -= new Vector2(0, 50);
-            JumpText.rectTransform.anchoredPosition -= new Vector2(0, 50);
-            MoveText.fontStyle |= FontStyles.Strikethrough;
-            JumpText.fontStyle |= FontStyles.Strikethrough;
-            // Show the Sprint text
-            JumpText.gameObject.SetActive(true);
-
-            
-
-            gameObject.SetActive(false);
-        }
-    }
-}
