--- SDP_Main/Assets/Scripts/UI/InGame/Toturial/Trigger2.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5b1b4deef57b5c748834b2c96e4aad2a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
