+++ SDP_Main/Assets/Scripts/UI/InGame/Toturial/TriggerScript.cs
@@ -1,31 +1,31 @@
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
-
 using UnityEngine.UI;
 using TMPro;
 
-public class Trigger1 : MonoBehaviour
+
+public class TriggerScript : MonoBehaviour
 {
     public TextMeshProUGUI MoveText;
     public TextMeshProUGUI JumpText;
 
-    public GameObject trigger2;
+
 
     private void OnTriggerEnter(Collider other)
     {
         Debug.Log("Trigger work");
         if (other.CompareTag("Player"))
         {
+            Debug.Log("Player work");
             //Move downward the tip text
             MoveText.rectTransform.anchoredPosition -= new Vector2(0, 50);
             MoveText.fontStyle |= FontStyles.Strikethrough;
             // Show the Jump text
             JumpText.gameObject.SetActive(true);
 
-            trigger2.SetActive(true);
-
-            gameObject.SetActive(false);
+        
+            Destroy(gameObject);
         }
     }
 }
